refactor(services): migrate external_events to handoff/
Move external_events.py to handoff/external_events.py:
- Handles external event recording for handoff system
- Used by handoff/service.py for CI status and PR comments

Updated import in handoff/service.py to use local path.
