+++ src/vibe3/services/handoff/external_events.py
